fixed loan class
